// src/typechain/out/AerodromeV2Adapter.ts
export type SwapAeroV2Struct = {
  amountToSend: BigNumberish;
  amountToReceive: BigNumberish;
  routes: IAerodromeRouter.RouteStruct[];
  deadline: BigNumberish;
};
// ...
export declare namespace IAerodromeRouter {
  export type RouteStruct = {
    from: AddressLike;
    to: AddressLike;
    stable: boolean;
    factory: AddressLike;
  };

  export type RouteStructOutput = [
    from: string,
    to: string,
    stable: boolean,
    factory: string
  ] & { from: string; to: string; stable: boolean; factory: string };
}