// src/typechain/out/UniswapV2Adapter.ts
export type SwapV2Struct = {
  amountToSend: BigNumberish;
  amountToReceive: BigNumberish;
  path: AddressLike[];
  deadline: BigNumberish;
};