// src/typechain/out/UniswapV3Adapter.ts
export type SwapV3Struct = {
  amountToSend: BigNumberish;
  amountToReceive: BigNumberish;
  path: BytesLike;
  deadline: BigNumberish;
};